[TDF] Move more utility functions to TDFUtils
