+++ tree/treeplayer/src/TDFUtils.cxx
@@ -1,0 +1,67 @@
+// Author: Enrico Guiraud, Danilo Piparo CERN  03/2017
+
+/*************************************************************************
+ * Copyright (C) 1995-2016, Rene Brun and Fons Rademakers.               *
+ * All rights reserved.                                                  *
+ *                                                                       *
+ * For the licensing terms see $ROOTSYS/LICENSE.                         *
+ * For the list of contributors see $ROOTSYS/README/CREDITS.             *
+ *************************************************************************/
+
+#include "ROOT/TDFUtils.hxx"
+#include "TROOT.h" // IsImplicitMTEnabled, GetImplicitMTPoolSize
+#include "TTree.h"
+
+#include <stdexcept>
+#include <string>
+
+namespace ROOT {
+namespace Internal {
+
+const char *ToConstCharPtr(const char *s)
+{
+   return s;
+}
+
+const char *ToConstCharPtr(const std::string s)
+{
+   return s.c_str();
+}
+
+unsigned int GetNSlots()
+{
+   unsigned int nSlots = 1;
+#ifdef R__USE_IMT
+   if (ROOT::IsImplicitMTEnabled()) nSlots = ROOT::GetImplicitMTPoolSize();
+#endif // R__USE_IMT
+   return nSlots;
+}
+
+void CheckTmpBranch(const std::string &branchName, TTree *treePtr)
+{
+   auto branch = treePtr->GetBranch(branchName.c_str());
+   if (branch != nullptr) {
+      auto msg = "branch \"" + branchName + "\" already present in TTree";
+      throw std::runtime_error(msg);
+   }
+}
+
+/// Returns local BranchNames or default BranchNames according to which one should be used
+const BranchNames_t &PickBranchNames(unsigned int nArgs, const BranchNames_t &bl, const BranchNames_t &defBl)
+{
+   bool useDefBl = false;
+   if (nArgs != bl.size()) {
+      if (bl.size() == 0 && nArgs == defBl.size()) {
+         useDefBl = true;
+      } else {
+         auto msg = "mismatch between number of filter arguments (" + std::to_string(nArgs) +
+                    ") and number of branches (" + std::to_string(bl.size() ? bl.size() : defBl.size()) + ")";
+         throw std::runtime_error(msg);
+      }
+   }
+
+   return useDefBl ? defBl : bl;
+}
+
+} // end NS Internal
+} // end NS ROOT
